requirements.txt
asgiref==3.5.2
black==22.3.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
drf-jwt==1.19.2
gunicorn==20.1.0
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
urllib3==1.26.9
wincertstore==0.2
